core: Lower the `WorkspaceStore` into the crate
Moving this is quite simple and adds another trait to the core crate.

Part of #6284
