+++ lib/src/workspace_store.rs
@@ -20,6 +20,8 @@ use std::io::Write as _;
 use std::path::Path;
 use std::path::PathBuf;
 
+pub use jj_core::workspace_store::WorkspaceStore;
+pub use jj_core::workspace_store::WorkspaceStoreError;
 use jj_lib::file_util::BadPathEncoding;
 use jj_lib::file_util::IoResultExt as _;
 use jj_lib::file_util::PathError;
@@ -35,39 +37,6 @@ use jj_lib::ref_name::WorkspaceName;
 use prost::Message as _;
 use tempfile::NamedTempFile;
 use thiserror::Error;
-
-/// Errors that can occur when interacting with a workspace store.
-#[derive(Error, Debug)]
-pub enum WorkspaceStoreError {
-    /// An unspecified error occurred.
-    #[error(transparent)]
-    Other(#[from] Box<dyn std::error::Error + Send + Sync>),
-}
-
-/// A storage backend for workspace metadata.
-pub trait WorkspaceStore: Send + Sync + Debug {
-    /// Returns the name of this workspace store implementation.
-    fn name(&self) -> &str;
-
-    /// Adds a workspace with the given name and path to the store.
-    fn add(&self, workspace_name: &WorkspaceName, path: &Path) -> Result<(), WorkspaceStoreError>;
-
-    /// Forgets the workspaces with the given names.
-    fn forget(&self, workspace_names: &[&WorkspaceName]) -> Result<(), WorkspaceStoreError>;
-
-    /// Renames a workspace from `old_name` to `new_name`.
-    fn rename(
-        &self,
-        old_name: &WorkspaceName,
-        new_name: &WorkspaceName,
-    ) -> Result<(), WorkspaceStoreError>;
-
-    /// Gets the path of the workspace with the given name, if it exists.
-    fn get_workspace_path(
-        &self,
-        workspace_name: &WorkspaceName,
-    ) -> Result<Option<PathBuf>, WorkspaceStoreError>;
-}
 
 /// Errors specific to the `SimpleWorkspaceStore` implementation.
 #[derive(Error, Debug)]
